CI only - tests, ruff, mypy, matrix. No release path, by design
spkproof is research, not a library, and is deliberately not published. The
release workflow and the PyPI contract check were removed rather than written,
so nothing here can publish it by accident.

Found and fixed: it could not run its own test suite from a fresh clone.

+++ src/spkproof/stats.py
@@ -43,4 +43,4 @@ def binom_sign_test(successes: int, n: int) -> float:
         return 1.0
     k = min(successes, n - successes)
     tail = sum(comb(n, i) for i in range(k + 1)) / (2 ** n)
-    return min(2.0 * tail, 1.0)
+    return float(min(2.0 * tail, 1.0))
